intelligence/requirements.txt
numpy
scikit-learn
torch
pytest
kafka-python
jsonschema